+++ wsxkj/src/main/java/com/zpj/materials/service/impl/GoodsTypeServiceImpl.java
@@ -3,6 +3,8 @@ package com.zpj.materials.service.impl;
 import com.zpj.common.BaseDao;
 import com.zpj.common.MyPage;
 import com.zpj.common.aop.Log;
+import com.zpj.materials.entity.Goods;
+import com.zpj.materials.entity.GoodsBrand;
 import com.zpj.materials.entity.GoodsType;
 import com.zpj.materials.service.GoodsTypeService;
 import com.zpj.sys.entity.LogInfo;
@@ -13,6 +15,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
@@ -63,17 +66,33 @@ public class GoodsTypeServiceImpl implements GoodsTypeService {
 	}
 
 	@Override
-	public void delInfo(String id, User user) {
+	public int delInfo(String id, User user) {
 		GoodsType temp=this.findById(id);
     	if(null!=temp){
-    		goodsTypeDao.delete(temp);
-    		LogInfo loginfo=new LogInfo();
-    		loginfo.setId(UUID.randomUUID().toString());
-    		loginfo.setUsername(user.getId());
-    		loginfo.setCreatetime(new Date());
-    		loginfo.setType("删除商品类型信息");
-    		loginfo.setDescription(temp.toString());
-    		logDao.add(loginfo);
+    		List  list =goodsTypeDao.findMapObjBySqlNoPage(" select id,typeId from jl_material_goods_brand_info where typeId='"+id+"'");
+    		if(list!=null&&list.size()>0){
+    			// 该分类有与之关联的品牌信息
+    			return 2;
+    		}else{
+    			List list1=goodsTypeDao.findMapObjBySqlNoPage(" select id,goodsType from jl_material_goods_info where goodsType ='"+id+"' ");
+    			if(list1!=null&&list1.size()>0){
+    				//该分类有与之关联的商品信息
+    				return 3;
+    			}else{
+    				goodsTypeDao.delete(temp);
+    				LogInfo loginfo=new LogInfo();
+    	    		loginfo.setId(UUID.randomUUID().toString());
+    	    		loginfo.setUsername(user.getId());
+    	    		loginfo.setCreatetime(new Date());
+    	    		loginfo.setType("删除商品类型信息");
+    	    		loginfo.setDescription(temp.toString());
+    	    		logDao.add(loginfo);
+    	    		return 1;
+    			}
+    			
+    		}
+    	}else{
+    		return 1;
     	}
 	}
 }
